feat: implement update-state event
Note that the state is not propagated to the other clients, after the
state had been succesfully updated.
In effect, the only purpose of the state is to give a new user its
initial state.

Its the duty of every client to send their changes directly to the other
clients (that's the idea of websockets)